refactor scripts, remove dead code
