+++ package.json
@@ -17,7 +17,7 @@
     "compile:unpacked": "yarn build && electron-builder build --config .electron-builder.config.js --dir --config.asar=false",
     "watch": "node watch.mjs",
     "prepare": "husky install",
-    "postinstall": "lerna run set-env-targets --scope=@gl/scripts"
+    "postinstall": "cross-env ELECTRON_RUN_AS_NODE=1 electron packages/scripts/bin/index.mjs update-electron-vendors"
   },
   "keywords": [],
   "author": {
