--- packages/automation/src/scoring/database/collections/default/index.ts
@@ -1,1 +1,0 @@
-export * from './v2'
