--- packages/automation/src/scoring/database/collections/default/v0.ts
@@ -1,51 +1,0 @@
-import {
-  RxJsonSchema,
-  toTypedRxJsonSchema,
-  ExtractDocumentTypeFromTypedRxJsonSchema,
-  MigrationStrategies,
-} from 'rxdb'
-
-const schemaLiteral = {
-  title: 'precalculation schema',
-  description: 'precalculated scores',
-  version: 0,
-  primaryKey: {
-    key: 'id',
-    fields: ['set', 'slot', 'main', 'sub'],
-    separator: '|',
-  },
-  type: 'object',
-  properties: {
-    id: {
-      type: 'string',
-      maxLength: 100,
-    },
-    set: {
-      type: 'string',
-    },
-    slot: {
-      type: 'string',
-    },
-    main: {
-      type: 'string',
-    },
-    sub: {
-      type: 'string',
-    },
-    popularity: {
-      type: 'number',
-      default: 0,
-    },
-  },
-  required: ['set', 'slot', 'main', 'sub'],
-} as const
-
-const schemaTyped = toTypedRxJsonSchema(schemaLiteral)
-
-export type RxDocType = ExtractDocumentTypeFromTypedRxJsonSchema<
-  typeof schemaTyped
->
-
-export const schema: RxJsonSchema<RxDocType> = schemaTyped
-
-export const migrationStrategies: MigrationStrategies = {}
