--- packages/automation/src/scoring/database/collections/index.ts
@@ -1,11 +1,0 @@
-import type { RxCollection, RxCollectionCreator } from 'rxdb'
-
-import * as Default from './default'
-
-export type Collections = {
-  default: RxCollection<Default.RxDocType>
-}
-
-export const collections: Record<keyof Collections, RxCollectionCreator> = {
-  default: Default.collection,
-}
