--- packages/automation/src/scoring/database/collections/types.ts
@@ -1,8 +1,0 @@
-import { OldRxCollection } from 'rxdb'
-
-type MaybePromise<T> = T | Promise<T>
-
-export type MigrationStrategy<OldDocData, DocData> = (
-  oldDocumentData: OldDocData,
-  oldRxCollection: OldRxCollection
-) => MaybePromise<DocData | null>
