--- packages/automation/src/scoring/database/index.ts
@@ -1,1 +1,0 @@
-export * from './database'
