+++ packages/automation/src/scoring/index.ts
@@ -1,4 +1,3 @@
 export * from './logic'
 export * from './types'
-export { getDatabase } from './database'
 export * from './const'
