+++ packages/scripts/package.json
@@ -5,13 +5,7 @@
   "author": "Lucas Pollice <lp1935@nyu.edu>",
   "homepage": "https://github.com/Jugbot/genshin-locker#readme",
   "license": "ISC",
-  "main": "src/index.ts",
-  "directories": {
-    "src": "src"
-  },
-  "files": [
-    "src"
-  ],
+  "bin": "bin/index.mjs",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Jugbot/genshin-locker.git"
@@ -22,13 +16,13 @@
     "lint:write": "eslint . --max-warnings=0 --ignore-path=\"../../.gitignore\" --fix",
     "lint": "eslint . --max-warnings=0 --ignore-path=\"../../.gitignore\"",
     "check-types": "tsc --noEmit",
-    "script": "vite-node --script src",
-    "set-env-targets": "cross-env ELECTRON_RUN_AS_NODE=1 electron src/update-electron-vendors.mjs"
+    "scripts": "scripts"
   },
   "bugs": {
     "url": "https://github.com/Jugbot/genshin-locker/issues"
   },
   "dependencies": {
+    "@nx/devkit": "^18.0.4",
     "@gl/automation": "*",
     "@gl/types": "*",
     "cli-progress": "^3.11.2",
@@ -42,7 +36,7 @@
     "electron": "^24.0.0",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.0",
-    "vite-node": "^0.33.0",
-    "vite": "4.2.1"
+    "vite": "4.2.1",
+    "vite-node": "^0.33.0"
   }
 }
